Move groups, static-model and types from base folder to a static folder
