+++ apps/backend-functions/src/notification.ts
@@ -58,7 +58,7 @@ import { appUrl, supportEmails } from './environments/environment';
 import { getReviewer } from '@blockframes/contract/negotiation/utils';
 // #7946 this may be reactivated later
 // import { createMailContract, MailContract } from '@blockframes/contract/contract/+state/contract.firestore';
-// import { movieCurrencies } from '@blockframes/utils/static-model';
+// import { movieCurrencies } from '@blockframes/model';
 
 // @TODO (#2848) forcing to festival since invitations to events are only on this one
 const eventAppKey: App = 'festival';
