+++ libs/model/src/lib/organisation.ts
@@ -1,6 +1,6 @@
 import { Location, createLocation } from '@blockframes/utils/common-interfaces/utility';
 import { OrgAppAccess, createOrgAppAccess, Module, App, getAllAppsExcept } from '@blockframes/utils/apps';
-import type { OrgActivity, OrganizationStatus } from './types';
+import type { OrgActivity, OrganizationStatus } from './static';
 import { createStorageFile, StorageFile, StorageVideo } from './media';
 import { DocumentMeta } from './meta';
 import { Timestamp } from './timestamp';
